+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 ## **WORK IN PROGRESS**
 
 * (theimo1221) Wled as iDimmable Lamp
+* (theimo1221) Persist Button Presses
 
 ## 1.1.2 (2022-09-10)